requirements.txt
# Trashformer Robot - Python Dependencies
# 
# RASPBERRY PI INSTALLATION NOTES:
# ================================
# 1. System packages (install first):
#    sudo apt-get update
#    sudo apt-get install -y python3-pip python3-dev python3-smbus i2c-tools
#    sudo apt-get install -y libatlas-base-dev libopenblas-dev libjasper-dev
#    sudo apt-get install -y libqtgui4 libqt4-test libhdf5-dev
#
# 2. Enable I2C:
#    sudo raspi-config
#    -> Interface Options -> I2C -> Enable
#
# 3. Install Python packages:
#    pip install -r requirements.txt
#
# 4. For OpenCV on older Pi models, use:
#    pip install opencv-python-headless  # Lighter version
#
# ============================================================================

# ============================================================================
# Configuration & Utilities
# ============================================================================
PyYAML>=6.0              # YAML configuration file parsing
python-dotenv>=0.19.0    # Environment variable management

# ============================================================================
# Hardware Interface
# ============================================================================
smbus2>=0.4.1            # I2C communication for PCA9685
RPi.GPIO>=0.7.1          # GPIO control (Raspberry Pi - only works on Pi)
pyserial>=3.5            # Serial communication for Sabertooth

# ============================================================================
# Computer Vision
# ============================================================================
opencv-python>=4.5.0     # Computer vision and camera interface
numpy>=1.21.0            # Numerical computing
Pillow>=9.0.0            # Image processing

# Machine Learning (optional - comment out if not using)
torch>=2.0.0             # PyTorch for deep learning
torchvision>=0.15.0      # PyTorch vision utilities
onnxruntime>=1.12.0      # ONNX model inference

# ============================================================================
# Sensors
# ============================================================================
adafruit-circuitpython-bno055>=5.0.0  # IMU sensor (if using Adafruit BNO055)
# Alternative IMU libraries (uncomment if needed)
# mpu6050-raspberrypi>=1.0.0
# adafruit-circuitpython-mpu6050>=1.0.0

# ============================================================================
# Input Devices
# ============================================================================
pygame>=2.1.0            # Gamepad/joystick support
keyboard>=0.13.5         # Keyboard input handling

# ============================================================================
# Data Processing & Math
# ============================================================================
scipy>=1.7.0             # Scientific computing
pandas>=1.3.0            # Data manipulation (for logging/analysis)

# ============================================================================
# Robotics & Control
# ============================================================================
# ROS support (optional - uncomment if using ROS)
# rospy>=1.15.0

# ============================================================================
# Testing
# ============================================================================
pytest>=7.0.0            # Testing framework
pytest-cov>=3.0.0        # Test coverage

# ============================================================================
# Development Tools
# ============================================================================
black>=22.0.0            # Code formatting
flake8>=4.0.0            # Code linting
mypy>=0.950              # Type checking

# ============================================================================
# Logging & Monitoring
# ============================================================================
colorlog>=6.6.0          # Colored logging output (optional)

# ============================================================================
# Visualization (optional)
# ============================================================================
matplotlib>=3.5.0        # Plotting and visualization